+++ Dev/Dev2.Data/MathOperations/FunctionEvaluator.cs
@@ -11,10 +11,10 @@
 using System;
 using System.Globalization;
 using Dev2.Common;
+using Dev2.Common.Interfaces.Diagnostics.Debug;
 using Infragistics.Calculations.CalcManager;
 using Infragistics.Calculations.Engine;
 using Warewolf.Resource.Errors;
-using Dev2.Common.Interfaces.Diagnostics.Debug;
 
 namespace Dev2.MathOperations
 {