+++ Dev/Dev2.Runtime.WebServer/Executor/ExecutionDto.cs
@@ -1,0 +1,36 @@
+/*
+*  Warewolf - Once bitten, there's no going back
+*  Copyright 2021 by Warewolf Ltd <alpha@warewolf.io>
+*  Licensed under GNU Affero General Public License 3.0 or later.
+*  Some rights reserved.
+*  Visit our website for more information <http://warewolf.io/>
+*  AUTHORS <http://warewolf.io/authors.php> , CONTRIBUTORS <http://warewolf.io/contributors.php>
+*  @license GNU Affero General Public License <http://www.gnu.org/licenses/agpl-3.0.html>
+*/
+
+using System;
+using Dev2.Communication;
+using Dev2.Data.TO;
+using Dev2.DataList.Contract;
+using Dev2.Interfaces;
+using Dev2.Runtime.WebServer.Executor;
+using Dev2.Runtime.WebServer.TransferObjects;
+using Warewolf.Data;
+
+namespace Dev2.Runtime.WebServer
+{
+    public class ExecutionDto : IExecutionDto
+    {
+        public WebRequestTO WebRequestTO { get; set; }
+        public string ServiceName { get; set; }
+        public string PayLoad { get; set; }
+        public IDSFDataObject DataObject { get; set; }
+        public EsbExecuteRequest Request { get; set; }
+        public Guid DataListIdGuid { get; set; }
+        public Guid WorkspaceID { get; set; }
+        public IWarewolfResource Resource { get; set; }
+        public DataListFormat DataListFormat { get; set; }
+        public Dev2JsonSerializer Serializer { get; set; }
+        public ErrorResultTO ErrorResultTO { get; set; }
+    }
+}