+++ Dev/Dev2.Server/ServerLifecycleManager.cs
@@ -30,6 +30,8 @@ using Dev2.Common.Interfaces.Wrappers;
 using System.Collections.Generic;
 using System.Management;
 using Dev2.Runtime.Interfaces;
+using Dev2.Studio.Utils;
+using System.Security.Claims;
 using System.Reflection;
 using System.Threading.Tasks;
 using Dev2.Activities;
@@ -260,7 +262,6 @@ namespace Dev2
 
                                 Stop(false, 0, true);
                             }
-
                             var logger = _loggerFactory.New(new JsonSerializer(), _webSocketPool);
                             LogWarewolfVersion(logger);
 #if DEBUG
@@ -434,11 +435,6 @@ namespace Dev2
                 _timer = null;
             }
 
-            if(_usageLogger != null)
-            {
-                _usageLogger.Dispose();
-            }
-
             if(_pulseLogger != null)
             {
                 _pulseLogger.Dispose();
@@ -448,7 +444,10 @@ namespace Dev2
             {
                 _pulseTracker.Dispose();
             }
-
+            if(_usageLogger != null)
+            {
+                _usageLogger.Dispose();
+            }
             if(_serverEnvironmentPreparer != null)
             {
                 _serverEnvironmentPreparer.Dispose();
@@ -487,6 +486,7 @@ namespace Dev2
             _writer.WriteLine("done.");
         }
 
+
         void LoadSubscriptionProvider()
         {
             _writer.Write("Loading subscription provider...  ");
@@ -497,6 +497,7 @@ namespace Dev2
                 _writer.WriteLine("done.");
             }
         }
+
 #if DEBUG
 
         static void SetAsStarted()