+++ Dev/Dev2.Studio/ViewModels/WorkSurface/WorkSurfaceContextViewModel.cs
@@ -39,7 +39,6 @@ using Dev2.Studio.Interfaces.Enums;
 using Warewolf.Studio.ViewModels;
 using Dev2.ViewModels;
 using Dev2.Common.Interfaces;
-using Dev2.Instrumentation;
 using Warewolf.Studio.Resources.Languages;
 
 namespace Dev2.Studio.ViewModels.WorkSurface
@@ -68,8 +67,6 @@ namespace Dev2.Studio.ViewModels.WorkSurface
         readonly IPopupController _popupController;
         readonly Action<IContextualResourceModel, bool, System.Action> _saveDialogAction;
         IResourceChangeHandlerFactory _resourceChangeHandlerFactory;
-
-        private readonly IApplicationTracker _applicationTracker;
 
         #endregion private fields
 
@@ -151,8 +148,6 @@ namespace Dev2.Studio.ViewModels.WorkSurface
             WorkSurfaceViewModel = workSurfaceViewModel ?? throw new ArgumentNullException(nameof(workSurfaceViewModel));
 
             _windowManager = CustomContainer.Get<IWindowManager>();
-
-            _applicationTracker = CustomContainer.Get<IApplicationTracker>();
 
             if (WorkSurfaceViewModel is IWorkflowDesignerViewModel model)
             {
@@ -324,10 +319,6 @@ namespace Dev2.Studio.ViewModels.WorkSurface
 
         public void Debug(IContextualResourceModel resourceModel, bool isDebug)
         {
-            if (_applicationTracker != null)
-            {
-                _applicationTracker.TrackEvent(TrackEventDebugOutput.EventCategory, TrackEventDebugOutput.Debug);
-            }
             if (resourceModel?.Environment == null || !resourceModel.Environment.IsConnected)
             {
                 return;
@@ -397,10 +388,6 @@ namespace Dev2.Studio.ViewModels.WorkSurface
 
         public void QuickViewInBrowser()
         {
-            if (_applicationTracker != null)
-            {
-                _applicationTracker.TrackEvent(TrackEventDebugOutput.EventCategory,TrackEventDebugOutput.F7Browser);
-            }
             if (!ContextualResourceModel.IsWorkflowSaved)
             {
                 var successfuleSave = Save(ContextualResourceModel, true);
@@ -430,10 +417,6 @@ namespace Dev2.Studio.ViewModels.WorkSurface
 
         public void QuickDebug()
         {
-            if (_applicationTracker != null)
-            {
-                _applicationTracker.TrackEvent(TrackEventDebugOutput.EventCategory,TrackEventDebugOutput.F6Debug);
-            }
             if (DebugOutputViewModel.IsProcessing)
             {
                 StopExecution();