+++ Dev/Warewolf.UIBindingTests.WorkflowDesigner/Workflow Designer.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
 *  Warewolf - Once bitten, there's no going back
 *  Copyright 2021 by Warewolf Ltd <alpha@warewolf.io>
 *  Licensed under GNU Affero General Public License 3.0 or later.
@@ -50,6 +50,7 @@ using Dev2.Core.Tests;
 namespace Warewolf.UIBindingTests.WorkflowDesigner
 {
     [TestClass]
+    [TestCategory("CannotParallelize")]
     public class Workflow_Designer
     {
         Mock<IShellViewModel> _shellViewModelMock;